refactor: simplify WorkerTokenDep — real dependency, drop login guard
Make get_worker_token a proper FastAPI dependency that reads from
request.app.state instead of a stub requiring override in lifespan
and test conftest. Remove the YAGNI custom /jwt/login route — the
random UUID password is sufficient protection.
